Use the basePath from the settings in the endpoint
This change ensures that a client of hapi-swagger is able to configure
the API endpoint (including the protocol and host part at the
beginning). Actually it is possible to set the "endpoint" setting, but
because this is used as a path to a route, you are not able to set a
URL.
I think to use the bathPath is actually what a user expects and the
previous behaviour was a bug.