disabling dropdown for a property via meta
Joi.allow() creates a valid list that results on dropdown in Swagger
docs UI. This is not expected in many cases and here we add
`disableDropdown` meta property for those.

One example that is used in tests is non-negative integers that could
not be edited in swagger UI forms:

    Joi.number().integer().positive().allow(0)

Only by adding `.meta({disableDropdown: true})` one could make this
editable as a form field.